Move u40 to the root.
